pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbus-mppsolar"
version = "1.0.0"
description = "Venus OS D-Bus service for MPP Solar inverters"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "mpp-solar>=0.15.0",
    "pyserial>=3.5",
    "dbus-python>=1.2.16",
    "gobject>=2.28.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.9",
]

[project.urls]
Homepage = "https://github.com/yourusername/dbus-mppsolar"
Repository = "https://github.com/yourusername/dbus-mppsolar"
Issues = "https://github.com/yourusername/dbus-mppsolar/issues"

[tool.setuptools]
packages = ["dbus_mppsolar"]

[tool.setuptools.package-dir]
"" = "dbus-mppsolar"

[tool.black]
line-length = 120
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 120